requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.5
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
astunparse==1.6.3
attrs==25.1.0
billiard==4.2.1
branca==0.8.1
celery==5.5.3
certifi==2024.12.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
Django==5.1.5
django-admin-soft-dashboard==1.0.26
django-ckeditor==6.7.2
django-colorfield==0.12.0
django-cors-headers==4.7.0
django-dbbackup==4.3.0
django-debug-toolbar==5.0.1
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-modeltranslation==0.19.12
django-mptt==0.16.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.9
environs==14.1.0
et_xmlfile==2.0.0
filelock==3.17.0
flatbuffers==25.2.10
folium==0.19.5
fonttools==4.56.0
frozenlist==1.5.0
gast==0.6.0
geographiclib==2.0
geopy==2.4.1
google-pasta==0.2.0
graphviz==0.21
grpcio==1.74.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.8
kombu==5.5.4
libclang==18.1.1
magic-filter==1.0.12
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.1.0
namex==0.1.0
numpy==2.1.3
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.51
propcache==0.2.1
protobuf==5.29.5
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rich==14.1.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
str2bool==1.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
virtualenv==20.29.1
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
xyzservices==2025.4.0
yarl==1.18.3
